feat: refactor config types
SO-106
